TraceDag/HopTestUtils: move to common
