+++ projects/batfish/src/test/java/org/batfish/dataplane/traceroute/BUILD
@@ -5,20 +5,6 @@ package(
     default_visibility = ["//visibility:public"],
 )
 
-java_library(
-    name = "testlib",
-    srcs = ["HopTestUtils.java"],
-    deps = [
-        "//projects/batfish",
-        "//projects/batfish-common-protocol:common",
-        "//projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/matchers",
-        "@maven//:com_google_code_findbugs_jsr305",
-        "@maven//:com_google_guava_guava",
-        "@maven//:junit_junit",
-        "@maven//:org_hamcrest_hamcrest",
-    ],
-)
-
 junit_tests(
     name = "tests",
     srcs = glob([
@@ -33,7 +19,6 @@ junit_tests(
         "//projects/batfish-common-protocol:common",
         "//projects/batfish-common-protocol:common_testlib",
         "//projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/matchers",
-        "//projects/batfish/src/test/java/org/batfish/dataplane/traceroute:testlib",
         "//projects/bdd",
         "@maven//:com_fasterxml_jackson_core_jackson_core",
         "@maven//:com_google_code_findbugs_jsr305",
